Make CAA checking more like DCV checking (#8491)
If the primary perspective CAA check returns an error, return early
rather than always kicking off the remote checks anyway.

As part of this change, rearrange the code in DoCAA to more closely
mirror the code in DoDCV. In particular, move the creation of the result
protobuf into a helper function which can be called at both the
early-return and late-return locations.